requirements.txt
openai-whisper==20230314
PySimpleGUI==4.60.5
python-docx==0.8.11
